Move ExprOp and AccumOp into smaller packages.
